Webgl / move wrapX logic to vector layer renderer
